// libraries/react/types/Direction.ts
/** Direction enum for tooltip positioning. */
export enum Direction {
  /** East position */
  E = 'E',
  /** North position */
  N = 'N',
  /** North-East position */
  NE = 'NE',
  /** North-West position */
  NW = 'NW',
  /** South position */
  S = 'S',
  /** South-East position */
  SE = 'SE',
  /** South-West position */
  SW = 'SW',
  /** West position */
  W = 'W',
}